Ported the lirc_rpi module to the Hummingboard
Only the Hummingboard i2ex has an onboard IR receiver.
Not all Hummingboard users are willing to solder one
on the other Hummingboard types as well. So, since
all Hummingboards also have GPIO capabilities, an IR
receiver connected to the GPIO should be able to work
as well. Also, none of the Hummingboards have an IR
sender, so having the possibility to support IR
senders in Lirc is welcome.

To add support for all these functions, the
Raspberry Pi lirc_rpi module was ported to the
Hummingboard.